Added auth

+++ api/customers/controllers.js
@@ -6,7 +6,7 @@ exports.fetchCustomers = async (req, res, next) => {
     return res.json(customers);
   } catch (error) {
     next(error);
-    return res.status(500).json({ message: error.message });
+    // return res.status(500).json({ message: error.message });
   }
 };
 
@@ -26,14 +26,18 @@ exports.fetchCustomer = async (customerId, next) => {
 
 exports.addCustomer = async (req, res, next) => {
   try {
-    const newCustomer = await Customer.create(req.body);
-    console.log(newCustomer);
-
-    return res
-      .status(201)
-      .json({ msg: "customer Added", payload: newCustomer });
+    if (req.user) {
+      const newCustomer = await Customer.create(req.body);
+      console.log(newCustomer);
+      return res
+        .status(201)
+        .json({ msg: "customer Added", payload: newCustomer });
+    } else {
+      return res
+        .status(401)
+        .json({ message: "you are not authorized to add a customer" });
+    }
   } catch (error) {
-    console.log("error", error);
     next(error);
   }
 };
@@ -41,26 +45,14 @@ exports.addCustomer = async (req, res, next) => {
 exports.deleteCustomer = async (req, res, next) => {
   try {
     const { customerId } = req.params;
-
-    // const customer = await Customer.findById(customerId);
-
-    await Customer.findByIdAndDelete(customerId);
-    res.status(204).end();
-    // if (String(req.user._id) === String(customer.owner._id)) {
-    //   const profileId = customer.owner.profile;
-    //   let
-    //   foundProfile = await Profile.findById(profileId);
-    // foundProfile.customers = foundProfile.customers.filter(
-    //   (customer) => JSON.stringify(customer) !== JSON.stringify(customerId)
-    // );
-    // await Profile.findByIdAndUpdate({ _id: profileId }, { ...foundProfile });
-    // await req.customer.remove();
-    // res.status(204).end();
-    // } else {
-    //   const error = new Error("you are not the owner of this customer!");
-    //   error.status = 401;
-    //   next(error);
-    // }
+    if (req.user) {
+      await Customer.findByIdAndDelete(customerId);
+      res.status(204).end();
+    } else {
+      res
+        .status(401)
+        .json({ message: "you are not authorized to delete a customer" });
+    }
   } catch (err) {
     next(err);
   }
@@ -70,22 +62,26 @@ exports.updateCustomer = async (req, res, next) => {
   try {
     const { customerId } = req.params;
 
-    // const id = req.trip._id;
     const customer = req.body;
-
-    const updatedCustomer = await Customer.findByIdAndUpdate(
-      customerId,
-      customer,
-      {
-        runValidators: true,
-        new: true,
-      }
-    );
-    console.log("+++++", updatedCustomer);
-    res.status(200).json({
-      msg: "customer Updated",
-      payload: updatedCustomer,
-    });
+    if (req.user) {
+      const updatedCustomer = await Customer.findByIdAndUpdate(
+        customerId,
+        customer,
+        {
+          runValidators: true,
+          new: true,
+        }
+      );
+      console.log("+++++", updatedCustomer);
+      res.status(200).json({
+        msg: "customer Updated",
+        payload: updatedCustomer,
+      });
+    } else {
+      res
+        .status(401)
+        .json({ message: "you are not authorized to update a customer" });
+    }
   } catch (err) {
     next(err);
   }
